Refactor flagship config
